feat(ui): L-shell step 5a — extract interactive Pane into a shared ESM module
Lift the per-workstream conversational Pane (chat + approval cards + composer +
voice + tool/media/MCP-error/verdict rendering) out of ui/static/app.js into a
new shared ES module shared_static/interactive.js, so BOTH deployments can
mount it: the standalone turnstone-server UI (its split-pane shell stays in
app.js and builds panes via window.InteractivePane) and — next, in step 5b —
the console L-shell over a node-proxied Tier-2 stream.

- Transport seam: a per-pane `base` prefix ("" local, "/node/{id}" proxied)
  threads through every request; createInteractivePane derives it from nodeId
  (the LOCALITY invariant — an interactive session lives on a cluster node).
- Host seam: the couplings only the surrounding shell knows (workstream name,
  focus, stream-error recovery, the --skip-permissions banner target, the MCP
  consent badge) route through an injected host adapter; the standalone shell
  supplies the real one (refetchWorkstreamsAndReassign + STANDALONE_HOST), the
  console factory a Tier-1 / no-op one.
- Embedded chrome: the standalone split-pane affordances (focus tracking,
  context menu, split/close buttons) are gated behind !embedded; the embedded
  path adds the INTERACTIVE persona tag.
- First legacy pane lifted into a real module: it exports the factory for the
  console shell's import and bridges window.* for the still-classic standalone
  shell (which builds panes only after the workstream fetch, so the deferred
  module has run). coordinator.js + the shared substrate stay classic.

The whole tool-output / media / MCP-error / verdict cluster moved with the Pane
(used only by it); the consent-BADGE subsystem stays in the standalone shell,
reached via host.onConsentDetected.

Verified: 104 JS-guard tests + a headless harness running the real module
(standalone + embedded chrome, node-proxy transport, lifecycle, zero errors),
node --check, prettier, ruff.
