+++ turnstone/ui/static/index.html
@@ -638,6 +638,7 @@
     <script src="/shared/katex-0.17.0/katex.min.js"></script>
     <script src="/shared/hljs-11.11.1/highlight.min.js"></script>
     <script src="/shared/renderer.js"></script>
+    <script type="module" src="/shared/interactive.js"></script>
     <script src="/static/app.js"></script>
   </body>
 </html>
